+++ src/ui/states/state-needs-auth-info.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Button from "../components/button.svelte";
+  import Button from "../atoms/button.svelte";
   import ModalFooter from "../layout/modal-footer.svelte";
   import ModalSection from "../layout/modal-section.svelte";
   import RowLayout from "../layout/row-layout.svelte";
